# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Code4Me V2 API",
    "description": "The complete API for Code4Me V2",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/project/create": {
      "post": {
        "tags": [
          "Project"
        ],
        "summary": "Create Project",
        "description": "Create a new project for an authenticated user.\n\nSteps:\n1. Validate the provided auth token to get session info.\n2. Verify the session token is valid.\n3. Create a new project in the database.\n4. Associate the project with the session and the user.\n5. Store project metadata in Redis.\n6. Return the project token as an HttpOnly cookie in the response.\n\nArgs:\n    project_to_create: Project data provided in the request body.\n    app: FastAPI dependency to access the app context.\n    auth_token: Auth token passed as a cookie for authentication.\n\nReturns:\n    JsonResponseWithStatus: JSON response with project token or error details.",
        "operationId": "create_project_api_project_create_post",
        "parameters": [
          {
            "name": "auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Auth Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProject"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProjectPostResponse"
                }
              }
            }
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProjectPostResponse"
                }
              }
            },
            "description": "Created"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/InvalidOrExpiredAuthToken"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidOrExpiredSessionToken"
                    }
                  ],
                  "title": "Response 401 Create Project Api Project Create Post"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/UserNotFoundError"
                    },
                    {
                      "$ref": "#/components/schemas/SessionNotFoundError"
                    }
                  ],
                  "title": "Response 404 Create Project Api Project Create Post"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProjectError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/user/create": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create User",
        "description": "Create a new user with email and password.\n\nArgs:\n    user_to_create (CreateUser):\n        User data from the request body.\n    app (App):\n        Application context with access to database and services.\n\nReturns:\n    JsonResponseWithStatus: JSON response indicating success or failure.\n\nFlow:\n    1. Check if a user already exists with the given email.\n    2. Insert the new user into the database.\n    3. Send a verification email via Celery.\n    4. Return HTTP 201 with the new user ID.",
        "operationId": "create_user_api_user_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserPostResponse"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserPostResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAlreadyExistsWithThisEmail"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateProject": {
        "properties": {
          "project_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Project Name",
            "description": "Project name"
          }
        },
        "type": "object",
        "required": [
          "project_name"
        ],
        "title": "CreateProject"
      },
      "CreateProjectError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Server failed to create a project!"
          }
        },
        "type": "object",
        "title": "CreateProjectError"
      },
      "CreateProjectPostResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Project created successfully."
          },
          "project_token": {
            "type": "string",
            "title": "Project Token",
            "description": "Created project token."
          }
        },
        "type": "object",
        "required": [
          "project_token"
        ],
        "title": "CreateProjectPostResponse"
      },
      "CreateUser": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "format": "email",
            "title": "Email",
            "description": "User's email address"
          },
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Name",
            "description": "User's full name"
          },
          "password": {
            "type": "string",
            "maxLength": 50,
            "minLength": 8,
            "format": "password",
            "title": "Password",
            "description": "User's password (will be hashed)",
            "writeOnly": true
          },
          "config_id": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Config Id",
            "description": "Configuration ID to use"
          }
        },
        "type": "object",
        "required": [
          "email",
          "name",
          "password",
          "config_id"
        ],
        "title": "CreateUser"
      },
      "CreateUserError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Server failed to create a new user!"
          }
        },
        "type": "object",
        "title": "CreateUserError"
      },
      "CreateUserPostResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "User created successfully. Please check your email for verification."
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id",
            "description": "Created user id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "CreateUserPostResponse"
      },
...
      "ErrorResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Error message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "InvalidOrExpiredAuthToken": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Authentication token not found! You are not authenticated or your token has expired. Login before you can perform this action."
          }
        },
        "type": "object",
        "title": "InvalidOrExpiredAuthToken"
      },
...
      "InvalidOrExpiredSessionToken": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Session not found! You are not authenticated or your session has expired. Login before you can perform this action."
          }
        },
        "type": "object",
        "title": "InvalidOrExpiredSessionToken"
      },
...
      "SessionNotFoundError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Session not found. You may need to log in again."
          }
        },
        "type": "object",
        "title": "SessionNotFoundError"
      },
...
      "UserAlreadyExistsWithThisEmail": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "User already exists with this email!"
          }
        },
        "type": "object",
        "title": "UserAlreadyExistsWithThisEmail"
      },
      "UserNotFoundError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "User not found"
          }
        },
        "type": "object",
        "title": "UserNotFoundError"
      },
...
    }
  }
}